+++ Software_Project.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.4.33213.308
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Software_Project", "Software_Project\Software_Project.csproj", "{AF5131BF-7168-4968-9D27-F81C462E4630}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ClassLibrary2", "ClassLibrary2\ClassLibrary2.csproj", "{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "POS", "POS\POS.csproj", "{48D0347A-E8E1-4598-AE46-7A9F3B81D729}"
 EndProject
 Global
@@ -25,14 +23,6 @@ Global
 		{AF5131BF-7168-4968-9D27-F81C462E4630}.Release|Any CPU.Build.0 = Release|Any CPU
 		{AF5131BF-7168-4968-9D27-F81C462E4630}.Release|x64.ActiveCfg = Release|x64
 		{AF5131BF-7168-4968-9D27-F81C462E4630}.Release|x64.Build.0 = Release|x64
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Debug|x64.Build.0 = Debug|Any CPU
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Release|Any CPU.Build.0 = Release|Any CPU
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Release|x64.ActiveCfg = Release|Any CPU
-		{35E4C2A6-A2F3-41C0-9E61-1013B6EAE616}.Release|x64.Build.0 = Release|Any CPU
 		{48D0347A-E8E1-4598-AE46-7A9F3B81D729}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{48D0347A-E8E1-4598-AE46-7A9F3B81D729}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{48D0347A-E8E1-4598-AE46-7A9F3B81D729}.Debug|x64.ActiveCfg = Debug|Any CPU
