+++ Software_Project/Choose.cs
@@ -13,7 +13,6 @@ using System.Windows.Forms;
 using System.Xml.Linq;
 using static Software_Project.Main_ui;
 using static System.Net.Mime.MediaTypeNames;
-using ClassLibrary2;
 
 namespace Software_Project
 {
