+++ Software_Project/table.cs
@@ -10,7 +10,6 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Windows.Markup;
-using ClassLibrary2;
 
 namespace Software_Project
 {
